requirements.txt
aiosmtplib==5.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
click==8.3.1
dnspython==2.8.0
email-validator==2.3.0
et_xmlfile==2.0.0
fastapi==0.133.0
h11==0.16.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
openpyxl==3.1.5
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
python-multipart==0.0.22
SQLAlchemy==2.0.47
sqlmodel==0.0.37
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.41.0
